Set high sensitivity to make file watching on OS X tolerable
Without this setting, it would take 5-10 seconds for an event to be
reported.

A caveat of this is that
`com.sun.nio.file.SensitivityWatchEventModifier` is an internal api
that might be removed in the future and require the `jdk.unsupported`
module to be added, see https://openjdk.java.net/jeps/260.

As of today it works in JDK 8 through 15.